+++ AtomicCore.BlockChain.BscscanAPI/JsonResult/Transaction/BscNormalTransactionJson.cs
@@ -54,7 +54,7 @@ namespace AtomicCore.BlockChain.BscscanAPI
         /// value
         /// </summary>
         [JsonProperty("value"), JsonConverter(typeof(BscBNBConverter))]
-        public BigInteger TxValue { get; set; }
+        public decimal TxValue { get; set; }
 
         /// <summary>
         /// contractAddress
